refactor(model): move discovery filter evaluation onto MatchFilter
A plugin announces its own discovery filters but could not test any of them:
MatchFilter is exported from pkg/model while the evaluation lived in the
agent's internal tree. A filter whose JSONPath is subtly wrong excludes nothing
and looks exactly like a resource that was not there, so the failure is silent
in the direction that leaves substrate exposed.

Move the evaluation to MatchFilter.Excludes, beside the type it belongs to.
pkg/model already depends on theory/jsonpath for its resource queries, so this
adds no dependency. ShouldFilterByMatchFilter stays as a thin wrapper over it,
leaving both call sites and the agent's own tests untouched, and the parser the
agent no longer needs goes with the code that used it.
